compositor|refactor: Rename three schema modules for their contents
`common.ts` is now `scalar-schemas.ts`, `plan-schema.ts` is now `plan-schemas.ts`, and `resolution-schemas.ts` is now `catalog-schemas.ts`, so each names what it declares rather than its audience or one of several subjects. The published API is unchanged.
