+++ packages/compositor/src/plan/__tests__/assertPlanIsConsistent.unit.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
 import { ConsistencyError } from '../../consistency/ConsistencyError.ts';
-import type { Plan } from '../../schemas/plan-schema.ts';
+import type { Plan } from '../../schemas/plan-schemas.ts';
 import { buildPlan } from '../../test-utils/buildPlan.ts';
 import { requireEntry } from '../../test-utils/requireEntry.ts';
 import { assertPlanIsConsistent, PlanConsistencyError } from '../assertPlanIsConsistent.ts';
