+++ packages/compositor/src/plan/checks/collectResolutions.ts
@@ -1,5 +1,5 @@
 import type { ResolutionAt } from '../../consistency/findResolutionOrderViolations.ts';
-import type { Plan } from '../../schemas/plan-schema.ts';
+import type { Plan } from '../../schemas/plan-schemas.ts';
 
 /** Collects each artifact's resolution, with the path prefix a violation against it is reported under. */
 export function collectResolutions(plan: Plan): ReadonlyArray<ResolutionAt> {
