+++ packages/compositor/src/plan/checks/findDuplicateFileKeys.ts
@@ -1,5 +1,5 @@
 import type { Violation } from '../../consistency/Violation.ts';
-import type { Plan } from '../../schemas/plan-schema.ts';
+import type { Plan } from '../../schemas/plan-schemas.ts';
 
 /**
  * Reports each destination two file entries both claim.
