+++ packages/compositor/src/plan/traversal.ts
@@ -1,8 +1,8 @@
 import { buildDependentsIndex } from '../graph/traversal.ts';
 import { appendTo } from '../portable/appendTo.ts';
-import type { ArtifactId } from '../schemas/common.ts';
 import type { FileEntry } from '../schemas/file-schemas.ts';
-import type { Plan } from '../schemas/plan-schema.ts';
+import type { Plan } from '../schemas/plan-schemas.ts';
+import type { ArtifactId } from '../schemas/scalar-schemas.ts';
 
 /** Reverse lookups over a plan's graph, answering the two "used by" directions the payload stores only forwards. */
 export interface TraversalIndex {
