+++ packages/compositor/src/portable/hash-content.ts
@@ -1,7 +1,7 @@
 import { Buffer } from 'node:buffer';
 import { createHash } from 'node:crypto';
 
-import type { Hash } from '../schemas/common.ts';
+import type { Hash } from '../schemas/scalar-schemas.ts';
 
 /** A digest of `data`, prefixed with the algorithm that produced it. */
 export function hashBytes(data: Uint8Array): Hash {
