+++ packages/compositor/src/resolution/__tests__/resolveCatalog.unit.test.ts
@@ -3,8 +3,8 @@ import path from 'node:path';
 
 import { describe, expect, it } from 'vitest';
 
-import type { CatalogEntry, ResolveKind, SourceSpec } from '../../schemas/resolution-schemas.ts';
-import { CATALOG_SCHEMA_VERSION, CatalogSchema } from '../../schemas/resolution-schemas.ts';
+import type { CatalogEntry, ResolveKind, SourceSpec } from '../../schemas/catalog-schemas.ts';
+import { CATALOG_SCHEMA_VERSION, CatalogSchema } from '../../schemas/catalog-schemas.ts';
 import { assertCatalogIsConsistent } from '../assertCatalogIsConsistent.ts';
 import { resolveCatalog } from '../resolveCatalog.ts';
 import { buildSource } from '../test-utils/buildSource.ts';
