+++ packages/compositor/src/resolution/checks/collectResolutions.ts
@@ -1,5 +1,5 @@
 import type { ResolutionAt } from '../../consistency/findResolutionOrderViolations.ts';
-import type { Catalog } from '../../schemas/resolution-schemas.ts';
+import type { Catalog } from '../../schemas/catalog-schemas.ts';
 
 /** Collects each entry's resolution, with the path prefix a violation against it is reported under. */
 export function collectResolutions(catalog: Catalog): ReadonlyArray<ResolutionAt> {
