+++ packages/compositor/src/resolution/checks/findIdDisagreements.ts
@@ -1,5 +1,5 @@
 import type { Violation } from '../../consistency/Violation.ts';
-import type { Catalog } from '../../schemas/resolution-schemas.ts';
+import type { Catalog } from '../../schemas/catalog-schemas.ts';
 import { composeArtifactId } from '../composeArtifactId.ts';
 
 /**
