+++ packages/compositor/src/samples/buildMinimalSample.ts
@@ -1,6 +1,6 @@
 import { hashUtf8 } from '../portable/hash-content.ts';
-import type { Plan } from '../schemas/plan-schema.ts';
-import { PLAN_SCHEMA_VERSION } from '../schemas/plan-schema.ts';
+import type { Plan } from '../schemas/plan-schemas.ts';
+import { PLAN_SCHEMA_VERSION } from '../schemas/plan-schemas.ts';
 
 const SKILL_BODY = '# Review\n\nRead the diff.\n';
