+++ packages/compositor/src/schemas/__tests__/catalog-schemas.unit.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import type { z } from 'zod';
 
-import { KindDescriptorSchema, SourceEntrySchema } from '../descriptor-schemas.ts';
 import {
   CATALOG_SCHEMA_VERSION,
   CatalogEntrySchema,
@@ -9,7 +8,8 @@ import {
   KindLayoutSchema,
   ResolveKindSchema,
   SourceSpecSchema,
-} from '../resolution-schemas.ts';
+} from '../catalog-schemas.ts';
+import { KindDescriptorSchema, SourceEntrySchema } from '../descriptor-schemas.ts';
 import { findIssuePaths } from '../test-utils/findIssuePaths.ts';
 
 const fileKind = {
