+++ packages/compositor/src/selection/selectArtifacts.ts
@@ -1,8 +1,8 @@
 import { compareStrings } from '../portable/compareStrings.ts';
-import type { ArtifactId, Id } from '../schemas/common.ts';
+import type { Catalog } from '../schemas/catalog-schemas.ts';
 import type { CompositorConfig, Selector } from '../schemas/config-schemas.ts';
 import type { Seed, SeedOrigin } from '../schemas/graph-schemas.ts';
-import type { Catalog } from '../schemas/resolution-schemas.ts';
+import type { ArtifactId, Id } from '../schemas/scalar-schemas.ts';
 import { buildCatalogIndex } from './buildCatalogIndex.ts';
 import { expandSelector } from './expandSelector.ts';
 import type { ConfigEntryRef, SelectionDiagnostic } from './SelectionDiagnostic.ts';
