+++ packages/compositor/src/selection/test-utils/buildCatalog.ts
@@ -1,7 +1,7 @@
 import { compareStrings } from '../../portable/compareStrings.ts';
 import { composeArtifactId } from '../../resolution/composeArtifactId.ts';
-import type { Catalog } from '../../schemas/resolution-schemas.ts';
-import { CATALOG_SCHEMA_VERSION } from '../../schemas/resolution-schemas.ts';
+import type { Catalog } from '../../schemas/catalog-schemas.ts';
+import { CATALOG_SCHEMA_VERSION } from '../../schemas/catalog-schemas.ts';
 
 /** One artifact the catalog carries, and the sources carrying a copy of it, highest precedence first. */
 export interface CatalogEntrySpec {
